Wooo! Images can load fine now
Issue was getting the dims of a loaded image

when you set img src (even if you're using a data blob) the full image
information isn't available straight away and you have to wait for an
onload callback to complete.

So...
I don't need to attach anything to the dom for it to be an avialble
resource. next up strip out resources div rubbish