Yeah
This is the start of a set of references you can make dependent on
something.

This is a bit specific and will need generalizing BUT it should hugely
simplify resource reloading on gl context being changed or a file change
or anything really.

Have to think about error conditions tho.

If I make the deref non lazy on depedency change I can call the new
resoure acquisition and check for error and just don't swap in new
resource. To play with